add all env vars to helm chart
